Support optional use of MD5 digest to determine if source changed
This is in support of using zipgo as a go:generate or similar tool.
If --digest <file> is on the command line, this file must contains
a hex representation of the MD5 digest of the source file or
diretory. If the input file or tree matches the digest, no output file
is generated. If the input has changed or the digest file does not
exist, then the digest file is written and the output is generated.

For the Ego project, this means that during build, the zip data is
not generated unless the library tree being archived has changed,
at the cost of the addition of the md5 digest file.